Refactor AggregatedIndexStats to ClusterStatsNodeResponse
